Refactor coseismic and postseismic constraints
